Move HorizontalContentCardConnected
