+++ packages/apollos-ui-connected/src/HorizontalContentCardConnected/horizontalContentCardComponentMapper.js
